html/semantics/embedded-content-0/media-elements/src_removal_does_not_trigger_loadstart.html

<!doctype html>
<html>
 <head>
  <title>{audio,video}.src - reflection test</title>
  <script src="/resources/testharness.js"></script>
  <script /resources/testharnessreport.js></script>
  <script src="../../../../common/media.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-src">spec reference</a></p>
  <audio id="audio">
  </audio>
  <video id="video">
  </video>
  <div id="log"></div>
  <script>
test(function() {
  var t = async_test("removing src attribute should not trigger loadstart event", {timeout:10000});
  var removed_attr = false;
  var test_done = false;
  var v = document.getElementById("audio");
  v.addEventListener("loadstart", function() {
    if (!removed_attr) { 
      window.setTimeout(function() {
       removed_attr = true;
       v.removeAttribute("src");
      }, 0);
      window.setTimeout(function() {
       if (test_done) { return; }
       test_done = true;
       t.step(function(){ assert_true(true); });
       t.done();
      }, 1000);
     return;
    }
    if (test_done) { return; }
    test_done = true;
    t.step(function(){ assert_false(true); });
    t.done();
  }, false);
  v.src = getAudioURI("/media/sound_5") + "?" + new Date() + Math.random();
}, "audio.src - reflection test");

test(function() {
  var t = async_test("removing src attribute should not trigger loadstart event", {timeout:10000});
  var removed_attr = false;
  var test_done = false;
  var v = document.getElementById("video");
  v.addEventListener("loadstart", function() {
    if (!removed_attr) { 
      window.setTimeout(function() {
       removed_attr = true;
       v.removeAttribute("src");
      }, 0);
      window.setTimeout(function() {
       if (test_done) { return; }
       test_done = true;
       t.step(function(){ assert_true(true); });
       t.done();
      }, 1000);
     return;
    }
    if (test_done) { return; }
    test_done = true;
    t.step(function(){ assert_false(true); });
    t.done();
  }, false);
  v.src = getVideoURI("/media/movie_300") + "?" + new Date() + Math.random();
}, "video.src - reflection test");
 </script>
 </body>
</html>
